refactor: 주석, 네이밍, 컨벤션 관련 작은 수정 사항
