+++ airbnb/src/index.tsx
@@ -7,8 +7,6 @@ if (process.env.NODE_ENV === 'development') {
   worker.start();
 }
 
-fetch('/').then(console.log);
-
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement,
 );
